apps/finsage-agent/pyproject.toml
[project]
name = "finsage-agent"
version = "0.1.0"
description = "FinSAGE agent service — FastAPI + Agent Framework + Fabric Data Agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-framework-core>=1.0.1",
    "agent-framework-foundry>=1.0.1",
    "agent-framework-openai>=1.0.1",
    "azure-identity>=1.25.3",
    "fastapi>=0.115.0",
    "openai>=2.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
